+++ backend/routes/user.routes.js
@@ -6,6 +6,7 @@ import {
   editUser,
   getUserById,
   deleteUser,
+  updatePassword
 } from "../controllers/user.controller.js" // Import controller functions
 
 const router = express.Router()
@@ -15,6 +16,7 @@ router.get("/:id", getUserById)
 router.post("/register", registerUser)
 router.post("/login", loginUser)
 router.put("/:id", editUser)
+router.put("/:id/password", updatePassword)
 router.delete("/:id", deleteUser)
 
 export default router