+++ frontend/src/pages/auth/AdminRoute.jsx
@@ -1,4 +1,3 @@
-import React from "react"
 import { Navigate, useParams } from "react-router-dom"
 
 const ProtectedAdminRoute = ({ children }) => {