+++ frontend/src/pages/auth/Login.jsx
@@ -1,8 +1,8 @@
 import { useState } from "react"
 import { useTranslation } from "react-i18next"
 import { Link, useParams, useNavigate } from "react-router-dom"
-import { Container, Row, Col, Card, Form, Button } from "react-bootstrap"
-import { handleLogin } from "@/utils/users"
+import { Container, Row, Col, Card, Form, Button, Alert } from "react-bootstrap"
+import { handleLogin } from "@/utils/auth"
 
 const Login = () => {
   const { t } = useTranslation()
@@ -12,10 +12,7 @@ const Login = () => {
   const [email, setEmail] = useState("")
   const [password, setPassword] = useState("")
   const [error, setError] = useState("")
-
-  const handleLoginClick = () => {
-    handleLogin(email, password, setError, navigate, lang)
-  }
+  const [message, setMessage] = useState(null)
 
   return (
     <Container fluid className="d-flex flex-column min-vh-100 p-0">
@@ -27,7 +24,12 @@ const Login = () => {
           >
             <Card.Body className="p-5 w-100 d-flex flex-column">
               <h3 className="mb-5 text-center">{t("login.login")}</h3>
-              {error && <p className="text-danger text-center">{error}</p>}
+
+              {message && <Alert variant={message.type}>{message.text}</Alert>}
+              {!message && error && (
+                <Alert variant={error.type}>{error.text}</Alert>
+              )}
+
               <Form>
                 <Form.Group className="mb-4" controlId="formEmail">
                   <Form.Control
@@ -38,7 +40,6 @@ const Login = () => {
                     onChange={(e) => setEmail(e.target.value)}
                   />
                 </Form.Group>
-
                 <Form.Group className="mb-4" controlId="formPassword">
                   <Form.Control
                     type="password"
@@ -48,12 +49,20 @@ const Login = () => {
                     onChange={(e) => setPassword(e.target.value)}
                   />
                 </Form.Group>
-
                 <Button
                   variant="primary"
                   size="lg"
                   className="w-100 mb-4"
-                  onClick={handleLoginClick}
+                  onClick={() =>
+                    handleLogin(
+                      email,
+                      password,
+                      setError,
+                      setMessage,
+                      navigate,
+                      lang
+                    )
+                  }
                 >
                   {t("login.login")}
                 </Button>